feat:우리의 추억 페이지 구성 구현
